requirements.txt
opencv-python
google-generativeai
gtts
SpeechRecognition
pyttsx3
pygame
Pillow
pyaudio
numpy
requests
